Use "very nomagic" in search()
Use the "\V" flag (very nomagic) when calling the search() function from
StartExplorer().  This causes all characters in the search pattern to be
treated literally, so there is no need to escape characters like "[" or
"~" that would normally have special meaning when the "magic" option is
set.  Since tildes no longer require escaping, the EscapeTilde()
function can be removed.