Refactor the mechanism to create a unique name for duplicate names
Previous mechanism only be able to compare the name of two buffers. If
there are more than two buffers with duplicate names, sometimes we still
can not get a unique name for each one.

New mechanism scans from the beginning of each buffer's path to find
signature parts which could uniquely identify the buffer and construct
a name string with these parts for each one.