Do not use 'IgnoreBuffer()' in building buffer dictionaries
At the time when BufNew or BufAdd is triggered, newly added buffer always
return unlisted and unmodifiable which makes 'IgnoreBuffer()' fail.

There is no way to identify whether a buffer is going to be a normal
buffer or not when it is newly created. So, we just build the buffer
dictionaries for every buffer that has a name. Due to condition checking
for buffers without a name has already been added in previous commits,
we just remove these checkings now. Usually, special/plugin buffer
always has a unique name which would't be a problem.