Start MBE on VimEnter/TabEnter event if auto-starting is set
VimEnter is triggered after BufEnter event in which 'AutoUpdate()' is
disabled, 'g:miniBufExplorerAutoStart' might not be working until next
BufEnter event, we should start the MBE on VimEnter or TabEnter.

We can not start MBE on the BufEnter event before VimEnter, because that
would mess up with vimdiff which will make MBE buffer as a diffed buffer.