Only set '&ut' when allowed and it is unchanged yet
Set '&ut' too low might break many things, it is up to the user.

If multiple files are given in the auguments list and the last one has
not been edited yet, issuing ':q!' suppose to quit anyway, but it would
fail if '&ut' is too low (e.g. 300ms).