Improve quiting if MBE is the last open window
If the last file of argulist has not been edited yet, executing `:q`
command would fail. Executing ':q' twice or ':q!' should quit anyway.
Since MBE is the last open window, we need to some how behave the same.
Quiting when executing ':q' twice has been implemented, but quiting when
executing ':q!' is hard to implement, because how can we know the command
that triggered this autocmd is ':q!'?