Only need to reset the '0 mark if MBE is the last window
If quiting vim fails due to some reason, we reset the '0 mark to the
last file, and abort all the other operations, this would let vim hanle
the rest.

The nesting of autocmds must been enabled on quit command to let other
plugin window be able to fire necessary operation.