Update MBE on BufDelete is necessary
Commit c31e8a6 prevents MBE from being updated for any deleted normal
buffer, because 'BufDelete' event is triggered before the deletion
actually takes place, at which moment the previous 'IsBufferIgnored()'
function will always return false which makes 'AutoUpdate()' never
called. As our 's:BufList' is always keeping updated, it should be
checked in 'IsBufferIgnored()' too.

Previous condition check is mainly to prevent updating MBE on the
deleting stage of a renamed beffer, but there is acctually no way of
telling whether that buffer is going to be deleted or renamed at this
moment. For a renamed buffer, later 'BufAdd' event will add it back
after all, so we should update MBE whenever a buffer gets deleted.